refactor: made sendtolives3 func reusable
