+++ packages/deployment-service/src/executables/deploy-from-store.ts
@@ -5,38 +5,8 @@ import { GetObjectOutput } from 'aws-sdk/clients/s3'
 import AdmZip from 'adm-zip'
 import path from 'path'
 import rimraf from 'rimraf'
-
-type SendToS3Params = {
-  filePath: string
-  prefix: string
-  buildLocation: string
-}
-
-type SendToLiveS3Func = (params: SendToS3Params) => Promise<void | never>
-
-const sendToLiveS3: SendToLiveS3Func = async ({
-  filePath,
-  prefix,
-  buildLocation,
-}: SendToS3Params): Promise<void | never> =>
-  new Promise<void>((resolve, reject) =>
-    s3Client.upload(
-      {
-        Bucket: process.env.DEPLOYMENT_LIVE_BUCKET_NAME as string,
-        Key: `${prefix}/${filePath.substring(buildLocation.length)}`,
-        Body: fs.readFileSync(filePath),
-        ACL: 'public-read',
-      },
-      (error) => {
-        if (error) {
-          console.error(error)
-          reject(error)
-        }
-
-        resolve()
-      },
-    ),
-  )
+import { sendToLiveS3, SendToLiveS3Func } from './deploy-to-live'
+import { recurseDir } from 'src/utils'
 
 const getFromVersionS3 = async (location: string): Promise<GetObjectOutput | never> =>
   new Promise<GetObjectOutput>((resolve, reject) =>
@@ -73,40 +43,6 @@ const deleteCurrentLiveVersion = (prefix: string): Promise<void | never> =>
       },
     ),
   )
-
-const recurseDir = async (
-  {
-    dir,
-    prefix,
-    buildLocation,
-  }: {
-    dir: string
-    prefix: string
-    buildLocation: string
-  },
-  callback: SendToLiveS3Func,
-) => {
-  const entries = fs.readdirSync(dir)
-  await Promise.all(
-    entries.map((name) => {
-      // [Promise<[Promise<[Promise<[] | void>] | void>] | void>] | void>... or Promise<void>
-      const filePath = path.join(dir, name)
-      const stat = fs.statSync(filePath)
-      if (stat.isFile()) {
-        return callback({ filePath, buildLocation, prefix })
-      } else if (stat.isDirectory()) {
-        return recurseDir(
-          {
-            dir: filePath,
-            prefix,
-            buildLocation,
-          },
-          callback,
-        )
-      }
-    }),
-  )
-}
 
 export const deployFromStore = async ({
   pipeline,
