+++ packages/deployment-service/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './ownership'
 export * from './resolve-customer'
 export * from './resolve-developer-id'
 export * from './project-dir'
+export * from './recurse-dir'
